LabRat's proto-type stand
STL files for the proto-type stand. Note.. a) Either use a 90 degree micro-usb adapter, or you may have to trim down the plastic on the micro-usb cable, in order to fit in the window detailed in these STL files. Alternatively, work to addd a larger hole for the micro-usb cable to feed through.